Refactor notifications setup
