+++ lib/src/log.dart
@@ -1,5 +1,6 @@
 import 'dart:developer' as developer;
 
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logging/logging.dart';
@@ -10,7 +11,7 @@ const _loggersToShowInTerminal = {
   'Socket',
 };
 
-void setupLogger() {
+void setupLogging() {
   if (kDebugMode) {
     Logger.root.level = Level.FINE;
     Logger.root.onRecord.listen((record) {
@@ -33,6 +34,16 @@ void setupLogger() {
         debugPrint('[${record.loggerName}] ${record.message}');
       }
     });
+  }
+
+  // Crashlytics
+  if (kReleaseMode) {
+    FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterFatalError;
+    // Pass all uncaught asynchronous errors that aren't handled by the Flutter framework to Crashlytics
+    PlatformDispatcher.instance.onError = (error, stack) {
+      FirebaseCrashlytics.instance.recordError(error, stack, fatal: true);
+      return true;
+    };
   }
 }
